pytest: store test fixtures in conftest.py
And move helpers to a separate module.

This is the idiomatic way. Fixtures defined in `conftest.py` are
automatically loaded for every test.

This approach is also adopted by WPT, c.f.
https://github.com/web-platform-tests/wpt/search?q=filename%3Aconftest.py

References:
- https://docs.pytest.org/en/6.2.x/fixture.html#conftest-py-sharing-fixtures-across-multiple-files
- https://stackoverflow.com/questions/34466027/in-pytest-what-is-the-use-of-conftest-py-files
